+++ codes/codes.go
@@ -1,76 +1,78 @@
 package codes
 
+type Code byte
+
 var (
-	PosFixedNumHigh byte = 0x7f
-	NegFixedNumLow  byte = 0xe0
+	PosFixedNumHigh Code = 0x7f
+	NegFixedNumLow  Code = 0xe0
 
-	Nil byte = 0xc0
+	Nil Code = 0xc0
 
-	False byte = 0xc2
-	True  byte = 0xc3
+	False Code = 0xc2
+	True  Code = 0xc3
 
-	Float  byte = 0xca
-	Double byte = 0xcb
+	Float  Code = 0xca
+	Double Code = 0xcb
 
-	Uint8  byte = 0xcc
-	Uint16 byte = 0xcd
-	Uint32 byte = 0xce
-	Uint64 byte = 0xcf
+	Uint8  Code = 0xcc
+	Uint16 Code = 0xcd
+	Uint32 Code = 0xce
+	Uint64 Code = 0xcf
 
-	Int8  byte = 0xd0
-	Int16 byte = 0xd1
-	Int32 byte = 0xd2
-	Int64 byte = 0xd3
+	Int8  Code = 0xd0
+	Int16 Code = 0xd1
+	Int32 Code = 0xd2
+	Int64 Code = 0xd3
 
-	FixedStrLow  byte = 0xa0
-	FixedStrHigh byte = 0xbf
-	FixedStrMask byte = 0x1f
-	Str8         byte = 0xd9
-	Str16        byte = 0xda
-	Str32        byte = 0xdb
+	FixedStrLow  Code = 0xa0
+	FixedStrHigh Code = 0xbf
+	FixedStrMask Code = 0x1f
+	Str8         Code = 0xd9
+	Str16        Code = 0xda
+	Str32        Code = 0xdb
 
-	Bin8  byte = 0xc4
-	Bin16 byte = 0xc5
-	Bin32 byte = 0xc6
+	Bin8  Code = 0xc4
+	Bin16 Code = 0xc5
+	Bin32 Code = 0xc6
 
-	FixedArrayLow  byte = 0x90
-	FixedArrayHigh byte = 0x9f
-	FixedArrayMask byte = 0xf
-	Array16        byte = 0xdc
-	Array32        byte = 0xdd
+	FixedArrayLow  Code = 0x90
+	FixedArrayHigh Code = 0x9f
+	FixedArrayMask Code = 0xf
+	Array16        Code = 0xdc
+	Array32        Code = 0xdd
 
-	FixedMapLow  byte = 0x80
-	FixedMapHigh byte = 0x8f
-	FixedMapMask byte = 0xf
-	Map16        byte = 0xde
-	Map32        byte = 0xdf
+	FixedMapLow  Code = 0x80
+	FixedMapHigh Code = 0x8f
+	FixedMapMask Code = 0xf
+	Map16        Code = 0xde
+	Map32        Code = 0xdf
 
-	FixExt1  byte = 0xd4
-	FixExt2  byte = 0xd5
-	FixExt4  byte = 0xd6
-	FixExt8  byte = 0xd7
-	FixExt16 byte = 0xd8
-	Ext8     byte = 0xc7
-	Ext16    byte = 0xc8
-	Ext32    byte = 0xc9
+	FixExt1  Code = 0xd4
+	FixExt2  Code = 0xd5
+	FixExt4  Code = 0xd6
+	FixExt8  Code = 0xd7
+	FixExt16 Code = 0xd8
+	Ext8     Code = 0xc7
+	Ext16    Code = 0xc8
+	Ext32    Code = 0xc9
 )
 
-func IsFixedNum(c byte) bool {
+func IsFixedNum(c Code) bool {
 	return c <= PosFixedNumHigh || c >= NegFixedNumLow
 }
 
-func IsFixedMap(c byte) bool {
+func IsFixedMap(c Code) bool {
 	return c >= FixedMapLow && c <= FixedMapHigh
 }
 
-func IsFixedArray(c byte) bool {
+func IsFixedArray(c Code) bool {
 	return c >= FixedArrayLow && c <= FixedArrayHigh
 }
 
-func IsFixedString(c byte) bool {
+func IsFixedString(c Code) bool {
 	return c >= FixedStrLow && c <= FixedStrHigh
 }
 
-func IsExt(c byte) bool {
+func IsExt(c Code) bool {
 	return (c >= FixExt1 && c <= FixExt16) || (c >= Ext8 && c <= Ext32)
 }