+++ decode_slice.go
@@ -12,14 +12,14 @@ const sliceElemsAllocLimit = 1e4
 var sliceStringPtrType = reflect.TypeOf((*[]string)(nil))
 
 func (d *Decoder) DecodeArrayLen() (int, error) {
-	c, err := d.readByte()
+	c, err := d.readCode()
 	if err != nil {
 		return 0, err
 	}
 	return d.arrayLen(c)
 }
 
-func (d *Decoder) arrayLen(c byte) (int, error) {
+func (d *Decoder) arrayLen(c codes.Code) (int, error) {
 	if c == codes.Nil {
 		return -1, nil
 	} else if c >= codes.FixedArrayLow && c <= codes.FixedArrayHigh {
@@ -148,14 +148,14 @@ func decodeArrayValue(d *Decoder, v reflect.Value) error {
 }
 
 func (d *Decoder) DecodeSlice() ([]interface{}, error) {
-	c, err := d.readByte()
+	c, err := d.readCode()
 	if err != nil {
 		return nil, err
 	}
 	return d.decodeSlice(c)
 }
 
-func (d *Decoder) decodeSlice(c byte) ([]interface{}, error) {
+func (d *Decoder) decodeSlice(c codes.Code) ([]interface{}, error) {
 	n, err := d.arrayLen(c)
 	if err != nil {
 		return nil, err
@@ -176,7 +176,7 @@ func (d *Decoder) decodeSlice(c byte) ([]interface{}, error) {
 	return s, nil
 }
 
-func (d *Decoder) skipSlice(c byte) error {
+func (d *Decoder) skipSlice(c codes.Code) error {
 	n, err := d.arrayLen(c)
 	if err != nil {
 		return err